+++ src/Factory.php
@@ -4,11 +4,14 @@ declare(strict_types = 1);
 namespace Innmind\FileWatch;
 
 use Innmind\Server\Control\Server\Processes;
-use Innmind\TimeWarp\Halt;
-use Innmind\TimeContinuum\Period;
+use Innmind\Time\{
+    Halt,
+    Period,
+};
 
 final class Factory
 {
+    #[\NoDiscard]
     public static function build(
         Processes $processes,
         Halt $halt,