+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 6.0.0 - 2026-02-07
+
+### Changed
+
+- Requires PHP `8.4`
+- Requires `innmind/server-control:~7.0`
+- Requires `innmind/time:~1.0`
+
 ## 5.0.0 - 2025-05-08
 
 ### Changed